Added runtime release/development system property which determines behavior of Browse buttons It gets set in the launch scripts and is handled by the isRelease method (which should be NPE-proof). If the JRD is a release build, Agent and DecompilerDependency Browse buttons should now point to locations within jrd's dir. Also moved the getJrdLocation method to the Directories class (seemed more logical) and fixed some JavaDoc comments there too.
