+++ runtime-decompiler/src/test/java/org/jrd/backend/decompiling/ExpandableUrlTest.java
@@ -10,6 +10,7 @@ import org.opentest4j.AssertionFailedError;
 import java.net.MalformedURLException;
 import java.net.URL;
 
+import static org.jrd.backend.data.Directories.getJrdLocation;
 import static org.jrd.backend.decompiling.ExpandableUrl.*;
 import static org.junit.jupiter.api.Assertions.*;
